Rename file and refactor to use subcomponents
